ci: latest toolchains everywhere; fold the bundle module in; release a dead instance's vnet ports
CI now tracks latest: go-version stable, tinygo 0.42.0, golangci-lint
latest, node lts/* (per-version pins go stale and nothing says so).

The routing-policy bundle loses its own go.mod + replace directive —
it only ever imported one in-tree package, so it is now a package of
the main module. Its unsafe wasi ABI (the reason a sub-module was
tolerable: go vet flags the standard pointer↔uintptr guest idiom)
moves behind the tinygo tag in abi_tinygo.go, with a native main stub;
the constraint tests run natively as before. bundle.wasm rebuilt with
tinygo 0.42.0 to match the new lane pin; parity test passes.

Desk resilience, from a live failure (visor wasm instance died after
~4h; its vnet ports stayed claimed, faking liveness and blocking a
rebind; its panic went only to an unscrolled xterm): bottle vnet
gains per-instance ownership + releaseOwner, the Go adapter tags
claims with SKYWIRE_EXEC_ID, and skywire-exec releases the dead
instance's claims and mirrors abnormal exits (code + stderr tail) to
the console so crashes are diagnosable from DevTools/CDP.
