+++ .github/workflows/android-app.yml
@@ -29,7 +29,7 @@ jobs:
 
       - uses: actions/setup-go@v6
         with:
-          go-version: '1.27.0'
+          go-version: 'stable'
           cache: true
           cache-dependency-path: go.sum
