+++ .github/workflows/codeql.yml
@@ -27,7 +27,7 @@ jobs:
 
       - uses: actions/setup-go@v7
         with:
-          go-version: '1.27.0'
+          go-version: 'stable'
           cache: true
           cache-dependency-path: go.sum
