+++ .github/workflows/deployment-health.yml
@@ -26,6 +26,6 @@ jobs:
       - uses: actions/checkout@v7
       - uses: actions/setup-go@v7
         with:
-          go-version: '1.27.0'
+          go-version: 'stable'
       - name: Probe dmsg-server wss fronts
         run: go run ./cmd/skywire cli mdisc check --url https://dmsgd.skywire.dev
