+++ .github/workflows/release.yml
@@ -9,7 +9,7 @@ on:
       - 'v*'
 
 env:
-  GO_VERSION: '1.27.0'
+  GO_VERSION: 'stable'
   MUSL_RELEASE: 'v1.3.29'
   MODULE: github.com/skycoin/skywire
